+++ papi-ui/src/components/kitty-avatar.tsx
