+++ papi-ui/src/components/kitty-card.tsx
