+++ papi-ui/src/components/marketplace-list.tsx
@@ -1,6 +1,6 @@
 import { Grid, Heading } from "@radix-ui/themes";
 import { MarketplaceKittyCard } from "./marketplace-kitty-card";
-import type { KittyForSale } from "./types";
+import type { KittyForSale } from "../types";
 
 interface Props {
   kittiesForSale: KittyForSale[];
